refactor: Move export jar related code under tasks folder
