+++ src/tasks/buildArtifact/GenerateJarExecutor.ts
@@ -7,8 +7,8 @@ import * as _ from "lodash";
 import { basename, dirname, extname, isAbsolute, join, normalize, relative } from "path";
 import { Disposable, ProgressLocation, QuickInputButtons, QuickPickItem, Uri, window, WorkspaceFolder } from "vscode";
 import { sendInfo } from "vscode-extension-telemetry-wrapper";
-import { Jdtls } from "../java/jdtls";
-import { INodeData } from "../java/nodeData";
+import { Jdtls } from "../../java/jdtls";
+import { INodeData } from "../../java/nodeData";
 import { IExportJarStepExecutor } from "./IExportJarStepExecutor";
 import { IClasspath, IStepMetadata } from "./IStepMetadata";
 import { createPickBox, ExportJarMessages, ExportJarStep, ExportJarTargets, getExtensionApi, toPosixPath } from "./utility";
